refactor: simplify creative routing and geometry helpers
